remove console.log and move reportWevVitals
